+++ build.sbt
@@ -38,11 +38,11 @@ libraryDependencies ++= Seq (
 )
 
 libraryDependencies += "org.apache.hadoop" % "hadoop-client" % "2.7.4"
-
 libraryDependencies += "io.spray" %%  "spray-json" % "1.3.3"
 libraryDependencies += "com.fasterxml.jackson.core" % "jackson-databind" % "2.9.0"
 libraryDependencies += "com.fasterxml.jackson.module" %% "jackson-module-scala" % "2.9.0"
 libraryDependencies += "io.jvm.uuid" %% "scala-uuid" % "0.2.3"
+libraryDependencies += "com.amazonaws" % "aws-java-sdk" % "1.11.226"
 
 dependencyOverrides ++= Set(
   "io.netty" %% "netty" % "3.7.0",
