+++ src/main/scala/org/marvin/manager/ArtifactSaver.scala
@@ -28,71 +28,3 @@ object ArtifactSaver {
   case class SaveToLocal(artifactName: String, protocol:String)
   case class SaveToRemote(artifactName: String, protocol:String)
 }
-
-class ArtifactSaver(metadata: EngineMetadata) extends Actor with ActorLogging {
-  var conf: Configuration = _
-
-  override def preStart() = {
-    log.info(s"${this.getClass().getCanonicalName} actor initialized...")
-    conf = new Configuration()
-
-    if (sys.env.get("HADOOP_CONF_DIR") != None){
-      val confFiles:List[File] = getListOfFiles(sys.env.get("HADOOP_CONF_DIR").mkString)
-
-      for(file <- confFiles){
-        log.info(s"Loading ${file.getAbsolutePath} file to hdfs client configuration ..")
-        conf.addResource(new FileInputStream(file))
-      }
-    }
-
-    conf.set("fs.defaultFS", metadata.hdfsHost)
-  }
-
-  def generatePaths(artifactName: String, protocol: String): Map[String, Path] = {
-    Map(
-      "localPath" -> new Path(s"${metadata.artifactsLocalPath}/${metadata.name}/$artifactName"),
-      "remotePath" -> new Path(s"${metadata.artifactsRemotePath}/${metadata.name}/${metadata.version}/$artifactName/$protocol")
-    )
-  }
-
-  def getListOfFiles(path: String): List[File] = {
-    val dir = new File(path)
-    val extensions = List("xml")
-    dir.listFiles.filter(_.isFile).toList.filter { file =>
-      extensions.exists(file.getName.endsWith(_))
-    }
-  }
-
-  override def receive: Receive = {
-    case SaveToLocal(artifactName, protocol) =>
-      log.info("Receive message and starting to working...")
-      val fs = FileSystem.get(conf)
-      val uris = generatePaths(artifactName, protocol)
-
-      log.info(s"Copying files from ${uris("remotePath")} to ${uris("localPath")}")
-
-      fs.copyToLocalFile(false, uris("remotePath"), uris("localPath"), false)
-      fs.close()
-
-      log.info(s"File ${uris("localPath")} saved!")
-
-      sender ! Done
-
-    case SaveToRemote(artifactName, protocol) =>
-      log.info("Receive message and starting to working...")
-      val fs = FileSystem.get(conf)
-      val uris = generatePaths(artifactName, protocol)
-
-      log.info(s"Copying files from ${uris("localPath")} to ${uris("remotePath")}")
-
-      fs.copyFromLocalFile(uris("localPath"), uris("remotePath"))
-      fs.close()
-
-      log.info(s"File ${uris("localPath")} saved!")
-
-      sender ! Done
-
-    case _ =>
-      log.warning("Received a bad format message...")
-  }
-}
